Cleaned up most of the softwipe warnings

+++ benchmark/src/tree.cpp
@@ -3,7 +3,8 @@
 #include <tree.hpp>
 
 static void BM_tree_constructor(benchmark::State &state) {
-  std::string filename = data_files_dna[data_keys[state.range(0)]].second;
+  std::string filename =
+      data_files_dna[data_keys[static_cast<size_t>(state.range(0))]].second;
   for (auto _ : state) {
     rooted_tree_t tree{filename};
   }
@@ -12,8 +13,9 @@ static void BM_tree_constructor(benchmark::State &state) {
 BENCHMARK(BM_tree_constructor)->Arg(0)->Arg(1);
 
 static void BM_tree_reroot(benchmark::State &state) {
-  rooted_tree_t tree{data_files_dna[data_keys[state.range(0)]].second};
-  auto rl = tree.root_location(state.range(1));
+  rooted_tree_t tree{
+      data_files_dna[data_keys[static_cast<size_t>(state.range(0))]].second};
+  auto rl = tree.root_location(static_cast<size_t>(state.range(1)));
   for (auto _ : state) {
     tree.root_by(rl);
   }
@@ -27,8 +29,9 @@ BENCHMARK(BM_tree_reroot)
     ->Args({1, 120});
 
 static void BM_tree_generate_operations(benchmark::State &state) {
-  rooted_tree_t tree{data_files_dna[data_keys[state.range(0)]].second};
-  auto rl = tree.root_location(state.range(1));
+  rooted_tree_t tree{
+      data_files_dna[data_keys[static_cast<size_t>(state.range(0))]].second};
+  auto rl = tree.root_location(static_cast<size_t>(state.range(1)));
   for (auto _ : state) {
     benchmark::DoNotOptimize(tree.generate_operations(rl));
   }
@@ -42,8 +45,9 @@ BENCHMARK(BM_tree_generate_operations)
     ->Args({1, 120});
 
 static void BM_tree_generate_derivative_operations(benchmark::State &state) {
-  rooted_tree_t tree{data_files_dna[data_keys[state.range(0)]].second};
-  auto rl = tree.root_location(state.range(1));
+  rooted_tree_t tree{
+      data_files_dna[data_keys[static_cast<size_t>(state.range(0))]].second};
+  auto rl = tree.root_location(static_cast<size_t>(state.range(1)));
   for (auto _ : state) {
     benchmark::DoNotOptimize(tree.generate_derivative_operations(rl));
   }
